Add Custom Reports to reviewboard
Add new reports to Dashboard that provide information that the system
doesnt provide right now. The new reports available are
i) Requests not Closed but with ShipIt
ii) Requests not reviewed by anyone
iii) Request not reviewed by me

The plan is to add more reports and provide an administration url that allows
users to pick and choose which reports they want to see in their installations'
dashboard. It is also planned to provide Aggregate reports .