Add support for selecting reports in dashboard
Provide a Configure menu in Extensions admin that allows a site administrator
to select which of the reports all users can see. Right now, its an all or
nothing, so its not specific to any users or user groups, but will bring it
in later if needed. The report dict is used to construct the names of the
reports so it dynamically expands depending on the number of reports that
get added to the system.

The settings are stored as part of registeredextension table and will be
available in the extension to display only appropriate reports.